Isolate seeded key pair generation
